Updated search pages to use new BaseDataTable.php
